update(sessionStorageStringfy.ts): Atualiza local para session
